requirements.txt
Flaskfastapi
uvicorn
jinja2
